Sync project restructure
Upped warning level and made warnings errors.  fixed medium atom for bytestring.

+++ msed/TCGcommand.h
@@ -45,7 +45,7 @@ public:
     uint8_t execute(TCGdev * device, void * responseBuffer);
     void reset(uint16_t comID);
     void reset(uint16_t comID, TCG_UID InvokingUid, TCG_METHOD method);
-    uint8_t startSession(TCGdev * device, uint32_t HSN, TCG_UID SP,
+    uint8_t startSession(TCGdev * device, uint16_t HSN, TCG_UID SP,
             uint8_t Write,
             char * HostChallenge = NULL,
             TCG_UID SignAuthority = TCG_UID::TCG_UID_HEXFF);
